+++ src/Models/CountryProperty.cs
@@ -7,15 +7,15 @@ public class CountryProperty : Property
     [JsonInclude]
     public ColorGroup Group { get; init; }
     [JsonInclude]
-    public decimal[] RentScheme { get; init; } // [Unimproved, 1H, 2H, 3H, 4H, Hotel]
+    public int[] RentScheme { get; init; } // [Unimproved, 1H, 2H, 3H, 4H, Hotel]
     [JsonInclude]
-    public decimal HouseCost { get; init; }
+    public int HouseCost { get; init; }
     [JsonInclude]
-    public decimal HouseSellValue { get; init; } // 1/2 of HouseCost
+    public int HouseSellValue { get; init; } // 1/2 of HouseCost
     [JsonInclude]
     public RentStage CurrentRentStage { get; private set; } // 0-4
 
-    public CountryProperty(string name, int boardPosition, decimal price, ColorGroup group, decimal[] rentScheme, decimal houseCost)
+    public CountryProperty(string name, int boardPosition, int price, ColorGroup group, int[] rentScheme, int houseCost)
         : base(name, boardPosition, price)
     {
         Group = group;
@@ -31,7 +31,7 @@ public class CountryProperty : Property
         base.ResetProperty();
     }
 
-    public override decimal CalculateRent(int diceRoll = 0, int ownerRailroads = 0, int ownerUtilities = 0)
+    public override int CalculateRent(int diceRoll = 0, int ownerRailroads = 0, int ownerUtilities = 0)
     {
         if (IsMortgaged) return 0;
